split runtime handles into one module per concern
