+++ python/mirage/runtime/handles/file_handle.py
@@ -14,81 +14,8 @@
 
 from collections.abc import Iterable
 from dataclasses import dataclass
-from typing import Generic, Literal, TypeVar
 
-FlushKind = Literal["append", "write"]
-
-# The lowest offset a handle has written at, before it writes anything.
-# A sentinel rather than None because every write takes the minimum of
-# it and the new offset, and "nothing yet" has to lose that comparison.
-NO_WRITE = 2**63 - 1
-
-T = TypeVar("T")
-
-
-def plan_flush(base_len: int, low_write: int,
-               buf: bytes | bytearray) -> tuple[FlushKind, bytes]:
-    """Decide what a closing whole-file buffer owes the mount.
-
-    Every encoder buffers a whole file and has to answer the same
-    question at close: did this handle only add to the end, or did it
-    rewrite what was already there? Only the first can travel as a
-    delta, and answering "write" always is what makes an append loop
-    quadratic.
-
-    Args:
-        base_len (int): length the file had when the handle opened.
-        low_write (int): lowest offset this handle wrote at, or the
-            NO_WRITE sentinel when it never wrote.
-        buf (bytes | bytearray): the handle's whole buffer.
-
-    Returns:
-        tuple[FlushKind, bytes]: ("append", tail) when the handle only
-        extended the file, else ("write", whole buffer).
-    """
-    if base_len > 0 and low_write >= base_len and len(buf) >= base_len:
-        return "append", bytes(buf[base_len:])
-    return "write", bytes(buf)
-
-
-@dataclass(frozen=True, slots=True)
-class OpenMode:
-    """What an fopen-style mode string says about a handle.
-
-    One vocabulary for every dialect that opens by mode: quickjs's
-    ``std.open`` and monty's ``path_open`` pass these strings verbatim,
-    and preview1's oflags/rights/fdflags translate onto the same five
-    facts in ``WasiFs.path_open``.
-
-    Args:
-        writable (bool): the handle may mutate its buffer (w, a, x, +).
-        truncate (bool): opening discards existing content (w).
-        append (bool): the position starts at the end (a).
-        create (bool): a missing file is created (w, a, x).
-        exclusive (bool): an existing file refuses the open (x).
-    """
-
-    writable: bool
-    truncate: bool
-    append: bool
-    create: bool
-    exclusive: bool
-
-
-def parse_mode(mode: str) -> OpenMode:
-    """Read an fopen-style mode string into its five facts.
-
-    Args:
-        mode (str): the mode as the guest spelled it (``r``, ``w+b``,
-            ``a``, ...); unknown letters are ignored, matching fopen.
-    """
-    return OpenMode(
-        writable=any(c in mode for c in "wax+"),
-        truncate="w" in mode,
-        append="a" in mode,
-        create=any(c in mode for c in "wax"),
-        exclusive="x" in mode,
-    )
+from mirage.runtime.handles.flush import NO_WRITE, FlushKind, plan_flush
 
 
 @dataclass(slots=True)
@@ -252,66 +179,3 @@ def merge_writes(base: bytes, writes: Iterable[tuple[int, bytes]]) -> bytes:
     for offset, chunk in writes:
         handle.pwrite(offset, chunk)
     return bytes(handle.buf)
-
-
-class FileTable(Generic[T]):
-    """Open handles by integer id.
-
-    The table every encoder used to hand-roll beside its handle type:
-    ids are dense from ``first_id`` upward and never reused within a
-    run. Generic because the entries differ (a wasi fd table holds
-    directory and stdio entries beside file handles, a FUSE core holds
-    its kernel-dialect handle), while the bookkeeping never does.
-
-    Args:
-        first_id (int): the first id handed out; wasi starts above its
-            three stdio fds and the preopen, FUSE starts at 1.
-    """
-
-    def __init__(self, first_id: int = 1) -> None:
-        self._next = first_id
-        self._entries: dict[int, T] = {}
-
-    def add(self, entry: T) -> int:
-        """Track an entry under a fresh id and return the id.
-
-        Args:
-            entry (T): the entry to track.
-        """
-        fd = self._next
-        self._next += 1
-        self._entries[fd] = entry
-        return fd
-
-    def get(self, fd: int) -> T | None:
-        """The entry under `fd`, or None.
-
-        Args:
-            fd (int): the id to look up.
-        """
-        return self._entries.get(fd)
-
-    def pop(self, fd: int) -> T | None:
-        """Remove and return the entry under `fd`, or None.
-
-        Args:
-            fd (int): the id to release.
-        """
-        return self._entries.pop(fd, None)
-
-    def set(self, fd: int, entry: T) -> None:
-        """Place an entry under an explicit id.
-
-        Args:
-            fd (int): the id to occupy (wasi seeds stdio at 0-2 and
-                renumbers onto guest-chosen ids).
-            entry (T): the entry to track.
-        """
-        self._entries[fd] = entry
-
-    def values(self) -> Iterable[T]:
-        """Every tracked entry."""
-        return self._entries.values()
-
-    def __contains__(self, fd: int) -> bool:
-        return fd in self._entries
