+++ typescript/packages/core/src/index.ts
@@ -541,7 +541,7 @@ export {
 } from './runtime/config.ts'
 export { CrossMountError, EvalError } from './runtime/errors.ts'
 export { RuntimeVFS, type VFSEntry, type VFSStat } from './runtime/vfs.ts'
-export { FileTable, mergeWrites, planFlush, type FlushKind } from './runtime/handles.ts'
+export { FileTable, mergeWrites, planFlush, type FlushKind } from './runtime/handles/index.ts'
 export {
   EVALUATOR,
   isEvaluator,
