+++ typescript/packages/core/src/runtime/handles/file_handle.ts
@@ -12,79 +12,7 @@
 // limitations under the License.
 // ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-export type FlushKind = 'append' | 'write'
-
-// The lowest offset a handle has written at, before it writes anything.
-// A sentinel rather than a null because every write takes the minimum of
-// it and the new offset, and "nothing yet" has to lose that comparison.
-export const NO_WRITE = Number.MAX_SAFE_INTEGER
-
-/**
- * Decide what a closing whole-file buffer owes the mount.
- *
- * Every encoder buffers a whole file and has to answer the same
- * question at close: did this handle only add to the end, or did it
- * rewrite what was already there? Only the first can travel as a
- * delta, and answering "write" always is what makes an append loop
- * quadratic.
- *
- * Args:
- *   baseLen: length the file had when the handle opened.
- *   lowWrite: lowest offset this handle wrote at, or the NO_WRITE
- *     sentinel when it never wrote.
- *   buf: the handle's whole buffer.
- *
- * Returns:
- *   ['append', tail] when the handle only extended the file, else
- *   ['write', whole buffer].
- */
-export function planFlush(
-  baseLen: number,
-  lowWrite: number,
-  buf: Uint8Array,
-): [FlushKind, Uint8Array] {
-  if (baseLen > 0 && lowWrite >= baseLen && buf.length >= baseLen) {
-    return ['append', buf.slice(baseLen)]
-  }
-  return ['write', buf.slice()]
-}
-
-/**
- * What an fopen-style mode string says about a handle.
- *
- * One vocabulary for every dialect that opens by mode: quickjs's
- * `std.open` passes these strings verbatim, and Python's preview1
- * oflags/rights/fdflags translate onto the same five facts.
- */
-export interface OpenMode {
-  /** The handle may mutate its buffer (w, a, x, +). */
-  writable: boolean
-  /** Opening discards existing content (w). */
-  truncate: boolean
-  /** The position starts at the end (a). */
-  append: boolean
-  /** A missing file is created (w, a, x). */
-  create: boolean
-  /** An existing file refuses the open (x). */
-  exclusive: boolean
-}
-
-/**
- * Read an fopen-style mode string into its five facts.
- *
- * Args:
- *   mode: the mode as the guest spelled it (`r`, `w+b`, `a`, ...);
- *     unknown letters are ignored, matching fopen.
- */
-export function parseMode(mode: string): OpenMode {
-  return {
-    writable: /[wax+]/.test(mode),
-    truncate: mode.includes('w'),
-    append: mode.includes('a'),
-    create: /[wax]/.test(mode),
-    exclusive: mode.includes('x'),
-  }
-}
+import { NO_WRITE, planFlush, type FlushKind } from './flush.ts'
 
 /**
  * One buffered whole-file handle.
@@ -236,56 +164,4 @@ export function mergeWrites(base: Uint8Array, writes: [number, Uint8Array][]): U
     handle.pwrite(offset, chunk)
   }
   return handle.buf
-}
-
-/**
- * Open handles by integer id.
- *
- * The table every encoder used to hand-roll beside its handle type:
- * ids are dense from `firstId` upward and never reused within a run.
- * Generic because the entries differ (a FUSE core holds its
- * kernel-dialect handle beside prefetched data), while the bookkeeping
- * never does. Mirrors Python's `FileTable`.
- */
-export class FileTable<T> {
-  private nextId: number
-  private readonly entries = new Map<number, T>()
-
-  constructor(firstId = 1) {
-    this.nextId = firstId
-  }
-
-  /** Track an entry under a fresh id and return the id. */
-  add(entry: T): number {
-    const fd = this.nextId++
-    this.entries.set(fd, entry)
-    return fd
-  }
-
-  /** The entry under `fd`, or undefined. */
-  get(fd: number): T | undefined {
-    return this.entries.get(fd)
-  }
-
-  /** Remove and return the entry under `fd`, or undefined. */
-  pop(fd: number): T | undefined {
-    const entry = this.entries.get(fd)
-    this.entries.delete(fd)
-    return entry
-  }
-
-  /** Place an entry under an explicit id. */
-  set(fd: number, entry: T): void {
-    this.entries.set(fd, entry)
-  }
-
-  /** Every tracked entry. */
-  values(): IterableIterator<T> {
-    return this.entries.values()
-  }
-
-  /** Whether an entry sits under `fd`. */
-  has(fd: number): boolean {
-    return this.entries.has(fd)
-  }
 }
