+++ typescript/packages/core/src/runtime/handles/mode.test.ts
@@ -1,0 +1,35 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import { describe, expect, it } from 'vitest'
+import { parseMode } from './mode.ts'
+
+describe('parseMode', () => {
+  it('reads the five facts', () => {
+    const read = parseMode('r')
+    expect(read.writable).toBe(false)
+    expect(read.create).toBe(false)
+    const update = parseMode('r+b')
+    expect(update.writable).toBe(true)
+    expect(update.truncate).toBe(false)
+    expect(update.create).toBe(false)
+    const write = parseMode('w')
+    expect(write.writable && write.truncate && write.create).toBe(true)
+    const append = parseMode('a')
+    expect(append.append).toBe(true)
+    expect(append.truncate).toBe(false)
+    const exclusive = parseMode('x')
+    expect(exclusive.exclusive && exclusive.create).toBe(true)
+  })
+})
